Move expressMiddleware to express4 deep import
While expressMiddleware has no runtime dependencies, it does rely on the
specific API of Express 4. One of the nice things about Express is that
its development had slowed and it was a very stable target, but all good
things must come to an end: Express 5 may be coming soon
https://github.com/expressjs/express/issues/4920

(I kid: the changes in Express 5, like support for proper error handling
with async functions, look great.)

So let's preemptively put this under `/express4` so that if the Express
5 middleware needs to be a bit different (if nothing else, the req/res
types in the context function may vary) that we'll have an obvious place
to put it.

Fixes #6587.
