+++ packages/server/src/index.ts
@@ -10,6 +10,3 @@ export {
   ForbiddenError,
   UserInputError,
 } from './errors.js';
-
-// TODO(AS4): consider moving to `@apollo/server/express`
-export { expressMiddleware } from './express/index.js';
