+++ packages/server/src/standalone/index.ts
@@ -5,7 +5,7 @@ import express from 'express';
 import http, { IncomingMessage, ServerResponse } from 'http';
 import type { ListenOptions } from 'net';
 import type { ApolloServer } from '../ApolloServer';
-import { expressMiddleware } from '../express/index.js';
+import { expressMiddleware } from '../express4/index.js';
 import type { BaseContext, ContextFunction } from '../externalTypes';
 import { ApolloServerPluginDrainHttpServer } from '../plugin/drainHttpServer/index.js';
 import { urlForHttpServer } from '../utils/urlForHttpServer.js';
